moved the McpUtils class and updated comments
